feat(dj): update commands
