+++ packages/dj/index.ts
@@ -2,10 +2,13 @@
 import { whereisPj } from "@darwinia/util";
 import yargs from "yargs";
 
-import cmdRecipe from "./src/recipe";
+import cmdBalance from "./src/balance";
+// import cmdBlock from "./src/block";
+import cmdCodec from "./src/codec";
 import cmdConfig from "./src/config";
 import cmdRelay from "./src/relay";
 import cmdTransfer from "./src/transfer";
+import cmdTx from "./src/tx";
 
 
 // main
@@ -19,11 +22,14 @@ import cmdTransfer from "./src/transfer";
     const _ = yargs
         .usage("dj <hello@darwinia.network>")
         .help("help").alias("help", "h")
-        .version("version", whereisPj().version).alias("version", "V")
-          .command(cmdRecipe)
+          .version("version", whereisPj().version).alias("version", "V")
+          .command(cmdBalance)
+          // .command(cmdBlock)
+          .command(cmdCodec)
           .command(cmdConfig)
           .command(cmdRelay)
           .command(cmdTransfer)
+          .command(cmdTx)
           .argv;
 
     // show help if no input
