+++ packages/dj/src/tx.ts
@@ -1,0 +1,15 @@
+import yargs from "yargs";
+import { API } from "@darwinia/api";
+import { log } from "@darwinia/util";
+
+const cmdTx: yargs.CommandModule = {
+    command: "tx <hash>",
+    describe: "Get tx by hash",
+    handler: async (args: yargs.Arguments) => {
+        const hash: number | string = (args.hash as string);
+        const res = await API.getExtrinsic(hash);
+        log.ox(JSON.stringify(res, null, 2));
+    },
+}
+
+export default cmdTx;
